# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Digital Credential Common Service API
  version: 1.0.0
  description: |
    Multi-tenant API providing a format-agnostic abstraction layer for Digital Credential operations.
    Normalizes issue/verify/hold patterns across credential formats (AnonCreds, SD-JWT, mDL, W3C VC) and
    back-end agents (Traction/ACA-Py, Credo Agent Service), exposing a single homogeneous REST API.

    This API is organized into three distinct surfaces, each targeting a different audience:

    | Surface | Audience | Auth |
    |---|---|---|
    | **Platform Administration** | Service operators | `platform-admin` role |
    | **Tenant Management** | Tenant administrators | `tenants:admin` scope |
    | **Digital Credential Operations** | Application developers / API consumers | Domain scopes (`credentials:offer`, `credentials:verify`, etc.) |

    ## Authentication
    All endpoints (except health checks and public discovery) require a Bearer JWT issued by the app's
    OIDC provider at `/oidc/token`. Tokens are obtained via:
    - **User login**: Authorization Code + PKCE flow (federated to Keycloak)
    - **Service-to-service**: `client_credentials` grant

    ## Async Operations
    Credential operations that require external party action return `202 Accepted` with an Operation
    resource. Poll `GET /operations/:id` or register webhooks for completion notifications.

    ## Pagination
...

    ## Rate Limiting
...

    ## Idempotency
...

    ## Error Responses
...

...

...

...
servers:
...

...
security:
  - BearerAuth: []
paths:
...
  /api/v1/tenants/{tenantId}/connections/{connectionId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: connectionId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getConnection
      summary: Get connection details
      tags: [Connections]
      responses:
        "200":
          description: Connection details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteConnection
      summary: Remove a connection
      tags: [Connections]
      responses:
        "204":
          description: Connection removed

  # ─── Operations ──────────────────────────────────────────────────────
...
  /api/v1/tenants/{tenantId}/audit-logs:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: listAuditLogs
      summary: Query audit log entries
      description: |
        Retrieve structured audit log records for the tenant.
        Audit logs track business actions (issue, verify, revoke, create, update, delete, login, token_grant)
        and are stored in PostgreSQL (separate from operational Loki logs).
      tags: [Audit Logs]
      parameters:
        - $ref: "#/components/parameters/Cursor"
        - $ref: "#/components/parameters/Limit"
        - name: action
          in: query
          schema:
            type: string
            enum: [issue, verify, hold, revoke, create, update, delete, login, token_grant]
        - name: actorId
          in: query
          schema:
            type: string
          description: Filter by actor (user ID or client ID)
        - name: resourceType
          in: query
          schema:
            type: string
          description: Filter by resource type (e.g., credential, connection, tenant)
        - name: resourceId
          in: query
          schema:
            type: string
            format: uuid
        - name: operationId
          in: query
          schema:
            type: string
            format: uuid
          description: Filter audit entries linked to a specific operation
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Paginated audit log entries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PaginatedResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/AuditLogEntry"
  /api/v1/tenants/{tenantId}/audit-logs/{auditLogId}:
    parameters:
      - $ref: "#/components/parameters/TenantId"
      - name: auditLogId
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getAuditLogEntry
      summary: Get audit log entry detail
      tags: [Audit Logs]
      responses:
        "200":
          description: Audit log entry with full metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditLogEntry"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/tenants/{tenantId}/audit-logs/export:
    parameters:
      - $ref: "#/components/parameters/TenantId"
    get:
      operationId: exportAuditLogs
      summary: Export audit logs as CSV
      description: Returns audit log entries matching the filters as a downloadable CSV file.
      tags: [Audit Logs]
      parameters:
        - name: action
          in: query
          schema:
            type: string
        - name: since
          in: query
          schema:
            type: string
            format: date-time
        - name: until
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
                example: 'attachment; filename="audit-logs-2025-01-15.csv"'

  # ─── Webhook Ingestion ───────────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: App-issued JWT obtained via /oidc/token (client_credentials or authorization_code grant)
  parameters:
    TenantId:
      name: tenantId
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: Tenant UUID
...
    Cursor:
      name: cursor
      in: query
      schema:
        type: string
      description: Opaque pagination cursor from a previous response
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Number of items per page
...
  responses:
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: RESOURCE_NOT_FOUND
              message: "Tenant not found"
              request_id: "req-uuid"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                  expected:
                    type: string
                  actual:
                    type: string
                  message:
                    type: string
            request_id:
              type: string
              format: uuid
    PaginatedResponse:
      type: object
      properties:
        data:
          type: array
          items: {}
        pagination:
          type: object
          properties:
            next_cursor:
              type: string
              nullable: true
            has_more:
              type: boolean
            total_count:
              type: integer
              description: Optional total count (may be omitted for performance)
...
    ConnectorType:
      type: string
      enum: [traction, credo]
...
    Connection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        external_connection_id:
          type: string
        their_label:
          type: string
          nullable: true
        their_did:
          type: string
          nullable: true
        state:
          $ref: "#/components/schemas/ConnectionState"
        connector_type:
          $ref: "#/components/schemas/ConnectorType"
        protocol:
          $ref: "#/components/schemas/ConnectionProtocol"
        metadata:
          type: object
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ConnectionState:
      type: string
      enum: [invited, requested, responded, active, completed, abandoned]
    ConnectionProtocol:
      type: string
      enum: [didcomm-v1, didcomm-v2, openid4vc]
...
    AuditLogEntry:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        actor_id:
          type: string
        actor_type:
          type: string
          enum: [user, system, client]
        action:
          type: string
          enum: [issue, verify, hold, revoke, create, update, delete, login, token_grant]
        resource_type:
          type: string
        resource_id:
          type: string
          format: uuid
        operation_id:
          type: string
          format: uuid
          nullable: true
        metadata:
          type: object
          description: Additional context (e.g., changed fields, error details)
        ip_address:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time

    # ─── Tenant Dead-letter Job ────────────────────────────────────
...